Adds shift multiply for scaling

+++ vitis/software/Baremetal/src/uz/uz_prng_halton/uz_prng_halton.c
@@ -64,7 +64,7 @@ static float get_uniform_float(uint32_t i_th_element, uint32_t base)
     while (i_th_element !=0U)
     {
         f = f / (float)base;
-        r = r + f * (i_th_element % base);
+        r = r + f * (float)(i_th_element % base);
         i_th_element = i_th_element / base;
     }
     return r;
